+++ ChangeLog.md
@@ -140,6 +140,7 @@ Nazara Engine:
 - Added ConvexCollider2D::GetVertices
 - Added SegmentCollider2D::GetThickness
 - Fixed vertices generation/render queue submit when using multiples materials on a Tilemap
+- It is now possible to prevent CompoundCollider2D to override individual colliders properties
 
 Nazara Development Kit:
 - Added ImageWidget (#139)